PARQUET-72: Update POM to use ASF parent.
Adds the org.apache:apache POM as the parent, which sets up the standard
Apache release configuration, including repositories,
distributionManagement, and a release profile. The existing Sonatype
release info is still present and the maven-release-plugin is configured
to use the Sonatype settings. To move to Apache, remove the
maven-release-pluging configuration and the Sonatype profile,
distributionManagement, and repositories.

Removes settings that are supplied by the parent, like encoding
properties and plugin versions. Overrides parent settings for compiler
source and target versions. Updates scm links.

Adds info for mailing lists and issue tracker.