PARQUET-109: Update NOTICE, add binary LICENSE.
This makes a minor change to NOTICE to match the maven-generated project
name (Apache Parquet => Apache Parquet Format (Incubating)). There
should be no more additions to NOTICE needed. The Thrift NOTICE has no
additions beyond the standard Apache boilerplate and SLF4J has no NOTICE
file.

This also adds a LICENSE file in the resources that is included in the
binary distribution. This LICENSE represents the contents of the binary
distribution and includes the QOS.ch BSD license for SLF4J. There are no
other licenses required. Although the LICENSE.txt in the thrift binary
that is shaded includes other licenses, the files that they apply to are
not included in either source or binary form because the shaded jar
contains only the Thrift Java library.

This also removes the Thrift NOTICE.txt and LICENSE.txt files from the
binary distribution. The parquet-format NOTICE and LICENSE files
represent the contents of the jar.