PARQUET-407: Incorrect delta-encoding example
https://issues.apache.org/jira/browse/PARQUET-407

The minimum and the number of bits are incorrect at delta encoding Example 2 In `Encodings.md`.
In the example,

```
Example 2

7, 5, 3, 1, 2, 3, 4, 5, the deltas would be

-2, -2, -2, 1, 1, 1, 1
The minimum is -2, so the relative deltas are:

0, 0, 0, 3, 3, 3, 3

The encoded data is

header: 8 (block size), 1 (miniblock count), 8 (value count), 7 (first value)

block 0 (minimum delta), 2 (bitwidth), 000000111111b (0,0,0,3,3,3 packed on 2 bits)
```

The minimum is -2 and the relative deltas are 0, 0, 0, 3, 3, 3, 3. So, this should be corrected as below:

```
block -2 (minimum delta), 2 (bitwidth), 00000011111111b (0,0,0,3,3,3,3 packed on 2 bits)
```

Author: socialpercon <socialpercon@gmail.com>

Closes #35 from socialpercon/master and squashes the following commits:

3d5886a [socialpercon] Change incorrect delta-encoding example