PARQUET-655: Fixes LogicalTypes.md link
The current LogicalTypes.md link in README.md points to the the file in the old https://github.com/Parquet/parquet-format repository.

This PR replaces the stale link with a relative path so that it always points to the file in the right repository.

Author: Cheng Lian <lian@databricks.com>

Closes #41 from liancheng/parquet-655-logical-types-link and squashes the following commits:

f98ddb7 [Cheng Lian] Fixes LogicalTypes.md link