PARQUET-686: Add ColumnOrders to FileMetaData.
This adds a list of ColumnOrder to FileMetaData. The list should include
one ColumnOrder for each column in the order in the same order as the
file's column chunks. Each ColumnOrder is a struct with an Order enum
(signed, unsigned, or custom) and an optional string to describe custom
orders.