PARQUET-974: Add a special ColumnOrder for testing
PR #46 introduced ColumnOrder with the limitation that a reader should
ignore stats for a column if the corresponding ColumnOrder in
FileMetaData contains an unknown value.

This change adds a special column order 'InvalidOrder' that can be used
to in tests and should not be used otherwise.

This change also fixes the paths in Makefile.